Stage3

+++ sources/views/contacts/activities-table.js
@@ -8,14 +8,16 @@ import TableView from "../tableView";
 
 export default class ActivitiesTable extends JetView {
 	config() {
-		const datatable = TableView("contact", {activityTypesDB});
+		const _ = this.app.getService("locale")._;
+
+		const datatable = TableView("contact", {activityTypesDB}, _);
 		const AddActivityBtn = {
 			view: "button",
 			type: "icon",
 			icon: "fas fa-plus-square",
-			width: 180,
+			width: 200,
 			height: 40,
-			label: "Add activity",
+			label: _("Add activity"),
 			css: "customBtn",
 			click: () => this.addActivity()
 		};
@@ -33,21 +35,9 @@ export default class ActivitiesTable extends JetView {
 			table.data.filter(obj => +obj.ContactID === +contactID, "", true);
 		});
 
-		this.on(table, "onItemClick", (id, e) => {
-			const editIcon = "far fa-edit editIcon table-icon";
-			const className = e.target.className;
-			if (editIcon === className) {
-				this.editIcon(e, id);
-			}
-		});
 
-		this.on(table, "onItemClick", (id, e) => {
-			const editIcon = "far fa-trash-alt deleteIcon table-icon";
-			const className = e.target.className;
-			if (editIcon === className) {
-				this.deleteIcon(e, id);
-			}
-		});
+		this.on(table, "onItemClick", (id, e) => this.onEditIcon(id, e));
+		this.on(table, "onItemClick", (id, e) => this.onDeleteIcon(id, e));
 	}
 
 	urlChange() {
@@ -63,24 +53,29 @@ export default class ActivitiesTable extends JetView {
 		});
 	}
 
+	onDeleteIcon(id, e) {
+		const _ = this.app.getService("locale")._;
 
-	deleteIcon(e, id) {
-		webix
-			.confirm({
-				title: "Delete",
-				text: "Do you want to delete this record? Deleting cannot be undone."
-			})
-			.then(() => {
-				activitiesDB.remove(id);
-			});
+		const editIcon = "far fa-trash-alt deleteIcon table-icon";
+		const className = e.target.className;
+		if (editIcon === className) {
+			webix
+				.confirm({
+					title: _("Delete"),
+					text: _("Do you want to delete this record? Deleting cannot be undone.")
+				})
+				.then(() => {
+					activitiesDB.remove(id);
+				});
+		}
 	}
 
-	editIcon(e, id) {
-		this._popup.showWindow({
-			id,
-			mode: "edit",
-			table: "contacts"
-		});
+	onEditIcon(id, e) {
+		const editIcon = "far fa-edit editIcon table-icon";
+		const className = e.target.className;
+		if (editIcon === className) {
+			this._popup.showWindow({id, mode: "edit", table: "contacts"});
+		}
 	}
 
 	addActivity() {
